refactor: 패키지 구조 변경 기존 코드가 레이어드 패키징 + 도메인 패키징 방식으로 섞인 것 같아 레이어드 패키징 구조로 변경하였습니다.
